requirements.txt
asgiref==3.8.1
boto3==1.35.19
botocore==1.35.19
dj-database-url==2.2.0
Django==5.1.1
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-flex-fields==1.0.2
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
